move API errors and util files to correct package (domain) since its files that would rarely be modified and deals with api
