--- app/src/main/java/com/example/tasky/common/data/util/Error.kt
@@ -1,3 +1,0 @@
-package com.example.tasky.common.data.util
-
-sealed interface Error
