+++ app/src/main/java/com/example/tasky/common/domain/error/Error.kt
@@ -1,0 +1,3 @@
+package com.example.tasky.common.domain.error
+
+sealed interface Error
