+++ app/src/main/java/com/example/tasky/feature_login/domain/repository/UserRemoteRepository.kt
@@ -1,7 +1,7 @@
 package com.example.tasky.feature_login.domain.repository
 
-import com.example.tasky.common.data.util.DataError
-import com.example.tasky.common.data.util.Result
+import com.example.tasky.common.domain.Result
+import com.example.tasky.common.domain.error.DataError
 import com.example.tasky.feature_login.data.model.LoginUserResponse
 
 interface UserRemoteRepository {
