+++ app/src/main/java/com/example/tasky/feature_login/presentation/AuthenticationViewState.kt
@@ -1,6 +1,6 @@
 package com.example.tasky.feature_login.presentation
 
-import com.example.tasky.common.data.util.DataError
+import com.example.tasky.common.domain.error.DataError
 
 open class AuthenticationViewState {
     object Success : AuthenticationViewState()
