+++ app/src/main/java/com/example/tasky/feature_login/presentation/LoginViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.common.data.EmailPatternValidatorImpl
-import com.example.tasky.common.data.util.Result
+import com.example.tasky.common.domain.Result
 import com.example.tasky.common.domain.isValidName
 import com.example.tasky.common.domain.isValidPassword
 import com.example.tasky.feature_login.domain.repository.UserRemoteRepository
